More optimizations

+++ arm9/source/gdi.h
@@ -49,7 +49,7 @@ class cGdi {
 
     u16 blendColors(u16 color, u16 dest, u16 src, u16 opacity);
 
-    u32 blendColors32(u16 color, u32 dest, u32 src, u16 opacity);
+    u32 blendColors32(u32 color, u32 dest, u32 src, u16 opacity);
 
     void fillRectBlend(u16 color1, u16 color2, s16 x, s16 y, u16 w, u16 h, GRAPHICS_ENGINE engine,
                        u16 opacity);
